--- Content.Server/Stories/Force/Components/PushOnCollideComponent.cs
@@ -1,16 +1,0 @@
-using Robust.Shared.Audio;
-
-namespace Content.Shared.Damage.Components;
-
-/// <summary>
-/// Applies stamina damage when colliding with an entity.
-/// </summary>
-[RegisterComponent]
-public sealed partial class PushOnCollideComponent : Component
-{
-    [ViewVariables(VVAccess.ReadWrite), DataField("damage")]
-    public float Damage = 55f;
-
-    [DataField("sound")]
-    public SoundSpecifier? Sound;
-}
