+++ Content.Shared/Stories/PushOnCollide/PushOnCollideComponent.cs
@@ -1,0 +1,8 @@
+using Robust.Shared.Audio;
+
+namespace Content.Shared.SpaceStories.Damage.Components;
+
+[RegisterComponent]
+public sealed partial class PushOnCollideComponent : Component
+{
+}
